requirements.txt
dspy
pydantic
requests
beautifulsoup4
pandas
python-dotenv
tqdm
openai
